refactor(translate-language): move language resolution into shared
`determineTargetLanguage` and the helpers around it decide which language a
translation targets. That decision is about to be made in main — the flow's
orchestration moves there so it can survive a window detach — while the renderer
still needs it to pre-check a translation before offering it (the PDF path
already uses it to refuse a same-language run).

Two consumers in two processes, so the module belongs in `shared`. It qualifies
without any change: `language.ts` had only type imports, no runtime dependency
of its own.

`UNKNOWN_LANG_CODE` travels with them — it is the value they compare against,
and every caller of one is a caller of the other.
