+++ src/renderer/hooks/translate/__tests__/useDetectLang.test.ts
@@ -1,11 +1,11 @@
 import { toast } from '@renderer/services/toast'
 import { parseTranslateLangCode } from '@shared/data/preference/preferenceTypes'
+import { UNKNOWN_LANG_CODE } from '@shared/utils/translateLanguage'
 import { mockUsePreference } from '@test-mocks/renderer/usePreference'
 import { mockRendererLoggerService } from '@test-mocks/RendererLoggerService'
 import { act, renderHook } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { UNKNOWN_LANG_CODE } from '../../../utils/translate'
 import {
   detectLanguageByFranc,
   detectLanguageByLLM,
