+++ src/renderer/pages/translate/components/TranslateLanguageBar.tsx
@@ -1,12 +1,12 @@
 import { Button, Combobox, type ComboboxOption, Tooltip } from '@cherrystudio/ui'
 import { useLanguages } from '@renderer/hooks/translate'
 import { cn } from '@renderer/utils/style'
-import { UNKNOWN_LANG_CODE } from '@renderer/utils/translate'
 import type {
   TranslateBidirectionalPair,
   TranslateLangCode,
   TranslateSourceLanguage
 } from '@shared/data/preference/preferenceTypes'
+import { UNKNOWN_LANG_CODE } from '@shared/utils/translateLanguage'
 import { ArrowLeftRight } from 'lucide-react'
 import type { FC } from 'react'
 import { useCallback, useMemo } from 'react'
