+++ src/renderer/utils/translate/index.ts
@@ -1,8 +1,2 @@
-export {
-  determineTargetLanguage,
-  getTargetLanguageForBidirectional,
-  pickBidirectionalTarget,
-  UNKNOWN_LANG_CODE
-} from './language'
 export { createInputScrollHandler, createOutputScrollHandler, handleScrollSync } from './scrollSync'
 export { createTranslateStreamId, translateText } from './translateText'
